+++ pmd-core/src/main/java/net/sourceforge/pmd/SourceCodeProcessor.java
@@ -222,7 +222,7 @@ public class SourceCodeProcessor {
             ctx.setLanguageVersion(forceLanguage);
         } else {
             // otherwise determine by file extension
-            LanguageVersion languageVersion = configuration.getLanguageVersionOfFile(ctx.getSourceCodeFilename());
+            LanguageVersion languageVersion = configuration.getLanguageVersionOfFile(ctx.getSourceCodeFile().toString());
             ctx.setLanguageVersion(languageVersion);
         }
     }